separate the classes into modules
