+++ src/index.js
@@ -1,93 +1,6 @@
 import './style.css';
 import '@fortawesome/free-solid-svg-icons';
-
-class Todo {
-  constructor(desc, completed, id) {
-    this.desc = desc;
-    this.completed = completed;
-    this.id = id;
-  }
-}
-class UI {
-  static displayTodos = (todos) => {
-    const ul = document.querySelector('.list');
-    ul.innerHTML = '';
-    todos.forEach((tod) => {
-      ul.innerHTML += `
-        <li id=${tod.id} class="item">
-          <div class='group'>
-            <input id=${tod.id} class='bdan' name='checkbox' type="checkbox" ${
-        tod.completed && 'checked'
-      } />
-            <p id=${tod.id} class='${
-        tod.completed && 'line'
-      }' contenteditable="true">${tod.desc}</p>
-          </div>
-          <i class="fa-solid fa-trash"></i>
-        </li>`;
-    });
-  };
-
-  static generateId = () => Math.floor((1 + Math.random()) * 0x10000)
-    .toString(16)
-    .substring(1);
-
-  static clearValue = () => {
-    document.querySelector('.input').value = '';
-  };
-
-  static addLineThroughToDos = ({ checked, id }) => {
-    if (checked) {
-      document.querySelectorAll('p').forEach((pTag) => {
-        if (pTag.id === id) pTag.classList.add('line');
-      });
-    } else {
-      document.querySelectorAll('p').forEach((pTag) => {
-        if (pTag.id === id) pTag.classList.remove('line');
-      });
-    }
-  };
-}
-
-class Store {
-  static getTodos = () => {
-    let todos;
-    if (!localStorage.getItem('todos')) todos = [];
-    else todos = JSON.parse(localStorage.getItem('todos'));
-    return todos;
-  };
-
-  static addTodo = (todo) => {
-    const todos = Store.getTodos();
-    todos.push(todo);
-    localStorage.setItem('todos', JSON.stringify(todos));
-    UI.displayTodos(todos);
-  };
-
-  static removeTodo = (id) => {
-    let todos = Store.getTodos();
-    todos = todos.filter((todo) => todo.id !== id);
-    localStorage.setItem('todos', JSON.stringify(todos));
-    UI.displayTodos(todos);
-  };
-
-  static changeStateofToDos = (todos, { checked, id }) => todos.map((todo) => {
-    if (!checked && id === todo.id) return { ...todo, completed: false };
-    if (checked && id === todo.id) return { ...todo, completed: true };
-    return todo;
-  });
-
-  static handelCheckBox = (checkbox) => {
-    if (checkbox.classList.contains('bdan')) {
-      let todos = Store.getTodos();
-      checkbox.addEventListener('change', (e) => {
-        UI.addLineThroughToDos(e.target);
-        todos = Store.changeStateofToDos(todos, e.target);
-        localStorage.setItem('todos', JSON.stringify(todos));
-      });
-    }
-  };
-}
+import { UI, Store, Todo } from './modules/index.js';
 
 // EVENTS
 document.addEventListener('DOMContentLoaded', () => {
