+++ src/modules/index.js
@@ -1,0 +1,5 @@
+export { default as Todo } from './todo.js';
+
+export { default as UI } from './ui.js';
+
+export { default as Store } from './store.js';
