+++ src/modules/todo.js
@@ -1,0 +1,9 @@
+class Todo {
+  constructor(desc, completed, id) {
+    this.desc = desc;
+    this.completed = completed;
+    this.id = id;
+  }
+}
+
+export default Todo;
